Changed to wrap the Popover
